Make `Client.update` require named parameters
I noticed it happened often that I used the microgrid id instead of
the dispatch id to identify the to-be-updated dispatch which of course
is wrong.

This should make it more obvious when using the API what kind of ID should
be passed.

Optimally we have distinct types for Microgrid ID and Dispatch ID though.

Signed-off-by: Mathias L. Baumann <mathias.baumann@frequenz.com>