Make optional arguments keyword-only
Both `Client.create()` and `Client.list() now take all optional
arguments as keyword-only arguments. This is to improve readability
and make calls less error prone, as these argument also have sometimes
the same time, making it hard to figure out which is which (for
example, `start_from` and `start_to` in `list()` or `active` and
`dry_run` in `create`).

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>